Allow wider range of sentiment values

+++ portfolio/src/main/java/com/google/sps/commentart/Comment.java
@@ -3,10 +3,10 @@ package com.google.sps.commentart;
 public class Comment {
   private String commenterName;
   private String commentMsg;
-  private int sentiment;
+  private float sentiment;
   private RepliesManager repliesManager;
 
-  public Comment(String commenterName, String commentMsg, int sentiment, RepliesManager repliesManager) {
+  public Comment(String commenterName, String commentMsg, float sentiment, RepliesManager repliesManager) {
     this.commenterName = commenterName;
     this.commentMsg = commentMsg;
     this.sentiment = sentiment;
